worker/requirements.txt
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ffmpeg-python==0.2.0
future==0.18.3
kombu==5.3.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
python-dateutil==2.8.2
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
